requirements.txt
Flask>=2.2.5
flask-cors==4.0.0
flask-limiter==3.5.0
python-dotenv==1.0.1
google-cloud-aiplatform==1.45.0
gunicorn==21.2.0
email-validator==2.1.0
python-dateutil==2.8.2
python-json-logger==2.0.7